+++ CHANGELOG.md
@@ -26,6 +26,7 @@ as necessary. Empty sections will not end in the release notes.
 - Add namespace validation for rename operation.
 - Namespace validation now correctly reports only one conflict when deleting a namespace that has
   children, whereas previously it reported one conflict for each child.
+- Make sqlText optional for IcebergView.
 
 ### Commits